Refactor charge type unit tests
- Move unit tests checking for specific charge_types to their own file.
- Remove any unused imports according to PEP8 standards.

[627]
